chore: replace MockMLSService in WireSyncEngine
