+++ wire-ios-sync-engine/Tests/Source/Synchronization/Strategies/CallingRequestStrategyTests.swift
@@ -17,6 +17,7 @@
 //
 
 import Foundation
+import WireDataModelSupport
 import WireRequestStrategy
 import WireSyncEngineSupport
 @testable import WireSyncEngine
